Add files via upload
The project contains an admin and user side. The admin side manages all the management like adding and managing categories, authors, books, issued books, and so on. The admin has an important role in the management of this online library system.

For the user section, the users can log in as a student if they already had an account, otherwise, he/she has to fill up the signup form to register as a student. The student can edit their owns profile and can go through details of issued books and their returning status. This project makes a convenient way for students and the library to keep the records of books, borrowing and returning of books.